small refactory
